+++ src/providers/bttv/BttvEmotes.cpp
@@ -66,7 +66,7 @@ namespace {
                 name,
                 ImageSet{Image::fromUrl(getEmoteLinkV3(id, "1x"), 1),
                          Image::fromUrl(getEmoteLinkV3(id, "2x"), 0.5),
-                         Image::fromUrl(getEmoteLinkV3(id, "3x"), 0.286)},
+                         Image::fromUrl(getEmoteLinkV3(id, "3x"), 0.25)},
                 Tooltip{name.string + "<br>Global BetterTTV Emote"},
                 Url{emoteLinkFormat.arg(id.string)},
             });